% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_56.R
\name{actual_vapour_pressure}
\alias{actual_vapour_pressure}
\title{Actual vapour pressure derived from relative humidity}
\usage{
actual_vapour_pressure(TN, TX, ...)
}
\arguments{
\item{TN}{numeric, minimum daily temperature (Celsius)}

\item{TX}{numeric, maximum daily temperature (Celsius)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Calculation of actual vapour pressure (kPa; reference: \href{http://www.fao.org/3/X0490E/x0490e00.htm#Contents}{FAO 56}).
}
\details{
Either minimum and maximum relative humidity (\code{UN} and \code{UX}) or mean relative humidity (\code{UG}) should be specified.
In case of usage \code{UN} and \code{UX} the actual vapour pressure is calculated by equation 17.
In absence of \code{UN} and \code{UX} the actual vapour pressure is calculated by equation 19 using \code{UG}.
}
\examples{
actual_vapour_pressure(TN = 18.0, TX = 25.0, UN = 54.0, UX = 82.0)
actual_vapour_pressure(TN = 18.0, TX = 25.0, UG = (54.0 + 82.0) / 2)
}
