% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_meteo}
\alias{analyse_meteo}
\title{Create overview of available observations from \href{https://www.knmi.nl/home}{KNMI} or \href{https://cds.climate.copernicus.eu}{AgERA5}}
\usage{
analyse_meteo(file, ...)
}
\arguments{
\item{file}{character string, filename of raw-data created by \code{download_meteo_KNMI} or \code{download_meteo_AgERA5}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Based on the data (\code{file}) an overview will be created with available observations present in the data.
}
\details{
Hourly data will be aggregated to daily values before plotting.
Optionally a selection of the data can be analysed by specifying either \code{variable} and/or the period using \code{tstart} and \code{tend}.
The size of the plot can be adjusted by specifying \code{width} (default 10.0) and/or \code{height} (default 16.0).
}
