% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_precipitation_KNMI}
\alias{analyse_precipitation_KNMI}
\title{Double-mass curve for precipitation from \href{https://www.knmi.nl/home}{KNMI} meteo station}
\usage{
analyse_precipitation_KNMI(dir_out, station, tstart, tend, ...)
}
\arguments{
\item{dir_out}{character string, name of directory}

\item{station}{character string, name(s) of meteo stations}

\item{tstart}{date, start of period}

\item{tend}{date, end of period}

\item{...}{further arguments passed to or from other methods}
}
\description{
Double-mass curve for precipitation from \href{https://www.knmi.nl/home}{KNMI} meteo station
}
\examples{
# analyse meteo from KNMI (do not run!)
#analyse_precipitation_KNMI(dir_out = "../temp_knmi",
# station = c("De Bilt", "Maastricht"), tstart = "1990-01-01", tend = "2010-12-31")
}
