% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{download_meteo_KNMI}
\alias{download_meteo_KNMI}
\title{Download observations from \href{https://www.knmi.nl/home}{KNMI} meteo station}
\usage{
download_meteo_KNMI(dir_out, station, variable, tstart, tend, ...)
}
\arguments{
\item{dir_out}{character string, name of directory}

\item{station}{character string, name(s) of meteo stations}

\item{variable}{character string, name(s) of observed variables}

\item{tstart}{date, start of period}

\item{tend}{date, end of period}

\item{...}{further arguments passed to or from other methods}
}
\description{
Download observations from \href{https://www.knmi.nl/home}{KNMI} meteo station
}
\examples{
# download meteo from KNMI (do not run!)
#download_meteo_KNMI(dir_out = "../temp_knmi",
# station = "de Bilt", variable = "RH", tstart = "1980-01-01", tend = "1980-12-31")
}
