% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{download_precipitation_KNMI}
\alias{download_precipitation_KNMI}
\title{Download observations from \href{https://www.knmi.nl/home}{KNMI} precipitation station}
\usage{
download_precipitation_KNMI(dir_out, station, tstart, tend, ...)
}
\arguments{
\item{dir_out}{character string, name of directory}

\item{station}{character string, name(s) of precipitation stations}

\item{tstart}{date, start of period}

\item{tend}{date, end of period}

\item{...}{further arguments passed to or from other methods}
}
\description{
Download observations from \href{https://www.knmi.nl/home}{KNMI} precipitation station
}
\examples{
# download precipitation from KNMI (do not run!)
#download_precipitation_KNMI(dir_out = "../temp_knmi",
# station = "de-Bilt", tstart = "1980-01-01", tend = "1980-12-31")
}
