% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_56.R
\name{extraterrestrial_radiation}
\alias{extraterrestrial_radiation}
\title{Extraterrestrial radiation (MJ m-2 d-1)}
\usage{
extraterrestrial_radiation(date, lat)
}
\arguments{
\item{date}{date (yyyy-mm-dd)}

\item{lat}{lattitude (degrees, minutes)}
}
\description{
Reference: \href{http://www.fao.org/3/X0490E/x0490e00.htm#Contents}{FAO 56}; equation 21
}
\examples{
extraterrestrial_radiation(date = "2001-07-06", lat = 50.48)
}
