% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{get_station_meteo_KNMI}
\alias{get_station_meteo_KNMI}
\title{Get info of meteo stations from \href{https://www.knmi.nl/home}{KNMI}}
\usage{
get_station_meteo_KNMI()
}
\description{
Get info of meteo stations from \href{https://www.knmi.nl/home}{KNMI}
}
\examples{
# select all meteo stations from KNMI
get_station_meteo_KNMI()
}
