% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{get_station_precipitation_KNMI}
\alias{get_station_precipitation_KNMI}
\title{Get info of precipitation stations from \href{https://www.knmi.nl/home}{KNMI}}
\usage{
get_station_precipitation_KNMI()
}
\description{
Get info of precipitation stations from \href{https://www.knmi.nl/home}{KNMI}
}
\examples{
# select all precipitation stations from KNMI
get_station_precipitation_KNMI()
}
