% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMItools.R
\docType{package}
\name{KNMItools-package}
\alias{KNMItools}
\alias{KNMItools-package}
\title{KNMItools}
\description{
R-scripts for downloading meterological and precipitation data for all \href{https://www.knmi.nl/home}{KNMI} stations.
}
\author{
\strong{Maintainer}: Martin Mulder \email{Martin2.Mulder@wur.nl} [copyright holder]

}
