% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_56.R
\name{net_longwave_radiation}
\alias{net_longwave_radiation}
\title{Net longwave radiation (MJ m-2 d-1)}
\usage{
net_longwave_radiation(Rs, Rso, Tmin, Tmax, ea)
}
\arguments{
\item{Rs}{global radiation (MJ m-2 d-1)}

\item{Rso}{clear sky solar radiation (MJ m-2 d-1)}

\item{Tmin}{minimum temperature (Celsius)}

\item{Tmax}{maximum temperature (Celsius)}

\item{ea}{actual vapour pressure (kPa)}
}
\description{
Reference: \href{http://www.fao.org/3/X0490E/x0490e00.htm#Contents}{FAO 56}; equation 39
}
\examples{
net_longwave_radiation(Rs = 22.07, Rso = 30.90, Tmin = 12.3, Tmax = 21.5, ea = 1.40824)
}
