% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_56.R
\name{net_radiation}
\alias{net_radiation}
\title{Net radiation (Mj m-2 d-1)}
\usage{
net_radiation(date, Rs, Tmin, Tmax, ea, lat, alt, albedo = 0.23)
}
\arguments{
\item{date}{date (yyyy-mm-dd)}

\item{Rs}{global radiation (MJ m-2 d-1)}

\item{Tmin}{minimum temperature (Celsius)}

\item{Tmax}{maximum temperature (Celsius)}

\item{ea}{vapour pressure (kPa)}

\item{lat}{lattitude (degrees, minutes)}

\item{alt}{altitude above sea level (m)}

\item{albedo}{albedo (-)}
}
\description{
Reference: \href{http://www.fao.org/3/X0490E/x0490e00.htm#Contents}{FAO 56}; equation 40
}
\examples{
net_radiation(date = "2001-07-06", Rs = 22.07, Tmin = 12.3,
 Tmax = 21.5, ea = 1.40824, lat = 50.48, alt = 100.0)
}
