% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{opt_station_meteo_KNMI}
\alias{opt_station_meteo_KNMI}
\title{Get possible meteo stations from \href{https://www.knmi.nl/home}{KNMI}}
\usage{
opt_station_meteo_KNMI()
}
\description{
Get possible meteo stations from \href{https://www.knmi.nl/home}{KNMI}
}
\examples{
# select all meteo stations from KNMI
opt_station_meteo_KNMI()
}
