% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{opt_station_precipitation_KNMI}
\alias{opt_station_precipitation_KNMI}
\title{Get possible precipitation stations from \href{https://www.knmi.nl/home}{KNMI}}
\usage{
opt_station_precipitation_KNMI()
}
\description{
Get possible precipitation stations from \href{https://www.knmi.nl/home}{KNMI}
}
\examples{
# select all precipitation stations from KNMI
opt_station_precipitation_KNMI()
}
