% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{opt_variable_meteo_KNMI}
\alias{opt_variable_meteo_KNMI}
\title{Get possible variables of meteo stations from \href{https://www.knmi.nl/home}{KNMI}}
\usage{
opt_variable_meteo_KNMI(topt)
}
\arguments{
\item{topt}{frequency of observations, options: 'day' or 'hour'}
}
\description{
Get possible variables of meteo stations from \href{https://www.knmi.nl/home}{KNMI}
}
\examples{
# select all variables of meteo stations from KNMI
opt_variable_meteo_KNMI(topt = "day")
}
