% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI.R
\name{opt_variable_precipitation_KNMI}
\alias{opt_variable_precipitation_KNMI}
\title{Get possible variables of precipitation stations from \href{https://www.knmi.nl/home}{KNMI}}
\usage{
opt_variable_precipitation_KNMI()
}
\description{
Get possible variables of precipitation stations from \href{https://www.knmi.nl/home}{KNMI}
}
\examples{
# select all variables of meteo stations from KNMI
opt_variable_precipitation_KNMI()
}
