% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_56.R
\name{reference_evapotranspiration}
\alias{reference_evapotranspiration}
\title{Reference evapotranspiration (mm d-1)}
\usage{
reference_evapotranspiration(date, Q, TN, TX, UN, UX, FG, lat, alt, altw)
}
\arguments{
\item{date}{date (yyyy-mm-dd)}

\item{Q}{global radiation (kJ m-2 d-1)}

\item{TN}{minimum temperature (Celsius)}

\item{TX}{maximum temperature (Celsius)}

\item{UN}{minimum relative humidity (\%)}

\item{UX}{maximum relative humidity (\%)}

\item{FG}{wind speed at height \code{altw} (m s-1)}

\item{lat}{lattitude (degrees, minutes)}

\item{alt}{altitude above sea level (m)}

\item{altw}{height at which \code{FG} was measured (m)}
}
\description{
Reference: \href{http://www.fao.org/3/X0490E/x0490e00.htm#Contents}{FAO 56}; equation 6
}
\examples{
reference_evapotranspiration(date = "2001-07-06", Q = 22070, TN = 12.3,
 TX = 21.5, UN = 63, UX = 84, FG = 2.778, lat = 50.48, alt = 100.0, altw = 10.0)
}
