% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_56.R
\name{saturation_vapour_pressure}
\alias{saturation_vapour_pressure}
\title{Estimation of saturation vapour pressure}
\usage{
saturation_vapour_pressure(T)
}
\arguments{
\item{T}{numeric, air temperature (Celsius)}
}
\description{
Reference: \href{http://www.fao.org/3/X0490E/x0490e00.htm#Contents}{FAO 56}; equation 11
}
